+++ cypher2sql/src/main/java/org/cytosm/cypher2sql/lowering/PopulateJoins.java
@@ -6,7 +6,6 @@ import org.cytosm.common.gtop.implementation.relational.TraversalHop;
 import org.cytosm.cypher2sql.lowering.exceptions.BugFound;
 import org.cytosm.cypher2sql.lowering.exceptions.Cypher2SqlException;
 import org.cytosm.cypher2sql.lowering.exceptions.Unreachable;
-import org.cytosm.cypher2sql.lowering.exceptions.TypeError;
 import org.cytosm.cypher2sql.lowering.sqltree.*;
 import org.cytosm.cypher2sql.lowering.sqltree.from.FromItem;
 import org.cytosm.cypher2sql.lowering.sqltree.join.BaseJoin;
@@ -14,9 +13,10 @@ import org.cytosm.cypher2sql.lowering.sqltree.join.InnerJoin;
 import org.cytosm.cypher2sql.lowering.sqltree.join.LeftJoin;
 import org.cytosm.cypher2sql.lowering.sqltree.visitor.Walk;
 import org.cytosm.cypher2sql.lowering.typeck.VarDependencies;
+import org.cytosm.cypher2sql.lowering.typeck.expr.Expr;
 import org.cytosm.cypher2sql.lowering.typeck.rel.Relationship;
 import org.cytosm.cypher2sql.lowering.typeck.var.*;
-import org.cytosm.cypher2sql.lowering.typeck.var.expr.ExprTree;
+import org.cytosm.cypher2sql.lowering.typeck.expr.ExprTree;
 
 import static org.cytosm.cypher2sql.lowering.exceptions.fns.LambdaExceptionUtil.rethrowConsumer;
