+++ cypher2sql/src/main/java/org/cytosm/cypher2sql/lowering/SelectTreeBuilder.java
@@ -4,8 +4,8 @@ import org.cytosm.cypher2sql.cypher.constexpr.ConstExpressionFolder;
 import org.cytosm.cypher2sql.lowering.sqltree.*;
 import org.cytosm.cypher2sql.lowering.typeck.ClauseId;
 import org.cytosm.cypher2sql.lowering.typeck.VarDependencies;
-import org.cytosm.cypher2sql.lowering.typeck.var.Expr;
-import org.cytosm.cypher2sql.lowering.typeck.var.expr.ExprTreeBuilder;
+import org.cytosm.cypher2sql.lowering.typeck.expr.Expr;
+import org.cytosm.cypher2sql.lowering.typeck.expr.ExprTreeBuilder;
 import org.cytosm.cypher2sql.cypher.ast.*;
 import org.cytosm.cypher2sql.cypher.ast.clause.*;
 import org.cytosm.cypher2sql.cypher.ast.clause.projection.*;
