+++ cypher2sql/src/main/java/org/cytosm/cypher2sql/lowering/typeck/constexpr/ConversionException.java
@@ -1,4 +1,4 @@
-package org.cytosm.cypher2sql.lowering.typeck.var.constexpr;
+package org.cytosm.cypher2sql.lowering.typeck.constexpr;
 
 /**
  */
