+++ cypher2sql/src/main/java/org/cytosm/cypher2sql/lowering/typeck/constexpr/UnknownOperation.java
@@ -1,4 +1,4 @@
-package org.cytosm.cypher2sql.lowering.typeck.var.constexpr;
+package org.cytosm.cypher2sql.lowering.typeck.constexpr;
 
 /**
  */
