+++ cypher2sql/src/main/java/org/cytosm/cypher2sql/lowering/typeck/expr/ExprTreeBuilder.java
@@ -1,10 +1,9 @@
-package org.cytosm.cypher2sql.lowering.typeck.var.expr;
+package org.cytosm.cypher2sql.lowering.typeck.expr;
 
 import org.cytosm.cypher2sql.lowering.typeck.AvailableVariables;
 import org.cytosm.cypher2sql.lowering.typeck.var.Var;
-import org.cytosm.cypher2sql.lowering.typeck.var.Expr;
 import org.cytosm.cypher2sql.cypher.visitor.Walk;
-import org.cytosm.cypher2sql.lowering.typeck.var.constexpr.ConstVal;
+import org.cytosm.cypher2sql.lowering.typeck.constexpr.ConstVal;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.apache.commons.lang3.tuple.Pair;
 import org.apache.log4j.Logger;
