+++ cypher2sql/src/main/java/org/cytosm/cypher2sql/lowering/typeck/rel/Relationship.java
@@ -3,10 +3,10 @@ package org.cytosm.cypher2sql.lowering.typeck.rel;
 import org.cytosm.cypher2sql.lowering.typeck.AvailableVariables;
 import org.cytosm.cypher2sql.lowering.typeck.NameProvider;
 import org.cytosm.cypher2sql.lowering.typeck.var.NodeVar;
-import org.cytosm.cypher2sql.lowering.typeck.var.Expr;
+import org.cytosm.cypher2sql.lowering.typeck.expr.Expr;
 import org.cytosm.cypher2sql.lowering.typeck.var.Var;
 import org.cytosm.cypher2sql.lowering.typeck.var.AliasVar;
-import org.cytosm.cypher2sql.lowering.typeck.var.expr.ExprTreeBuilder;
+import org.cytosm.cypher2sql.lowering.typeck.expr.ExprTreeBuilder;
 import org.cytosm.cypher2sql.cypher.ast.clause.match.pattern.NodePattern;
 import org.cytosm.cypher2sql.cypher.ast.clause.match.pattern.RelationshipChain;
 import org.cytosm.cypher2sql.cypher.ast.clause.match.pattern.RelationshipPattern.SemanticDirection;
